# PERF: concatenation of MultiIndexed objects (MultiIndex.append)