@phofl 2023-07-12
@lukemanley could you add a test like in the linked dask issue?